+++ Sources/PippinAdapters/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [Unreleased]
 
+### Changed
+
+- Obsoleted `CrudViewControllerUITableViewDelegate.crudViewController(crudViewController:otherEditActionsFor:)` and introduced `CrudViewControllerUITableViewDelegate.crudViewController(crudViewController:otherSwipeActionsFor:)` to provide the equivalent functionality with the new iOS function, as `UITableViewRowAction` was deprecated in iOS 12.
+
 ## [5.2.1] 2020-08-08
 
 ### Fixed